requirements.txt
openai==0.27.2
python_telegram_bot==20.0
